requirements.txt
Django==3.2.23
djangorestframework==3.14.0
django-cors-headers==4.0.0
channels==3.0.5
channels-redis==3.4.1
daphne==3.0.2
redis==4.5.1
psycopg2-binary==2.9.5
elasticsearch==7.17.9
django-elasticsearch-dsl==7.3.0
django-elasticsearch-dsl-drf==0.20.0
django-crispy-forms==1.14.0
django-js-asset==2.1.0
crispy-bootstrap4==2022.1
django-ckeditor==6.5.1
Pillow==9.5.0
django-allauth==0.54.0
django-allauth-2fa==0.11.1
django-cleanup==6.0.0
python-decouple==3.8
django-environ==0.11.2
django-debug-toolbar==3.8.1
pytest-django==4.5.2
pytest==7.3.1
gunicorn==20.1.0
uvicorn==0.20.0
cryptography==3.4.7
PyJWT==2.6.0
python-magic==0.4.27
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
urllib3==1.26.15
simplejson==3.17.6
agora-token-builder==1.0.0
python-magic-bin==0.4.14
python-dotenv==1.0.0
django-redis==5.2.0
whitenoise==6.4.0
structlog==22.3.0
django-validators==1.0.1
drf-yasg==1.21.5
celery==5.2.7
django-celery-beat==2.5.0
django-celery-results==2.5.0
sentry-sdk==1.25.1
django-anymail==9.1
social-auth-app-django==5.2.0
django-storages==1.13.2
django-compressor==4.3
django-pipeline==1.6.14
django-health-check==3.16.1
